requirements.txt
chrisapp
tensorflow==1.12.0
pypng
numpy
Pillow
keras
matplotlib
scipy
scikit-image
opencv-python-headless
